Finished adding in routes
Finished cleaning up routes from app.js to standalone routes.
