+++ views/staff.hbs
@@ -222,7 +222,7 @@
 
         let getStaff = () => {
             document.getElementById('password-table').innerHTML = '';
-            fetch('/staff-list')
+            fetch('/staff/list')
                 .then(response => response.json())
                 .then(data => {
                     {{!-- console.log(data) --}}
